Team linue ups now put in to own component with the match day info button
