+++ src/components/MatchDay/index.js
@@ -1,6 +1,5 @@
 import { connect } from "react-redux";
 import MatchDay from "./MatchDay";
-import { getInfo } from "../../data/actions/state";
 
 let mapStateToProps = (state) => {
     return {
@@ -8,10 +7,4 @@ let mapStateToProps = (state) => {
     };
 };
 
-let mapDispatchToProps = (dispatch) => {
-    return {
-        handleMatchDay: () => dispatch(getInfo()),
-    };
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(MatchDay);
+export default connect(mapStateToProps)(MatchDay);
